fix methods to get free/used logical devices if high drive letter not existent
The binary string of the drive_bitmask is only so long as necesarry. If a
drive Z exists than it has 26 entries, but if the highest used drive is
C there are only 3 (0b100). This will ignore much free drives.
To workaround this the returned value will be increased by 2**26 - a
theoretical drive after Z to include all drives until Z.